NYC-crash-data/pyproject.toml
[project]
name = "NYC_crash_data"
requires-python = ">=3.10,<3.15"
version = "0.1.0"
dependencies = [
    "dagster==1.13.0",
    "dagster-duckdb>=0.29.0",
    "duckdb>=1.5.2",
    "openmeteo-requests>=1.7.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "requests-cache>=1.3.1",
    "retry-requests>=2.0.0",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
    "ruff>=0.15.10",
    "pyright>=1.1.408",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }
packages = ["src/NYC_crash_data"]

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "NYC_crash_data"
registry_modules = [
    "NYC_crash_data.components.*",
]
